add e2e test sutie
