+++ src/internal/tester/tester.go
@@ -4,6 +4,7 @@ import (
 	"reflect"
 	"runtime"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -23,4 +24,19 @@ func AreSameFunc(t *testing.T, expect, have any) {
 				Pointer(),
 		).Name(),
 	)
+}
+
+// LogTimeOfTest logs the test name and the time that it was run.
+func LogTimeOfTest(t *testing.T) string {
+	now := time.Now().UTC().Format(time.RFC3339Nano)
+	name := t.Name()
+
+	if name == "" {
+		t.Logf("Test run at %s.", now)
+		return now
+	}
+
+	t.Logf("%s run at %s", name, now)
+
+	return now
 }
